feat: MessagePack encoding parity for modified-OTLP sketch hot path (PR I)
Adds MessagePack as a parallel encoding option alongside the
protobuf-encoded sketchlib `*State` path that PR B and PR C wired up
for all five modified-OTLP `Metric.data` sketch variants. The new
encoding is a cross-language contract between DataCollector (Go,
sketchlib-go) and ASAPQuery-backend (Rust, sketch-core) — both sides
serialize the same cross-language sketch-core wire struct via
MessagePack. This gives sketchlib-go a practical alternative to
prost-encoded sketchlib proto without touching existing paths.

## Proto

Adds two new enum variants to each of the five per-sketch encoding
enums in the vendored `opentelemetry.proto.metrics.v1`:
  * `*_ENCODING_MSGPACK         = 3`
  * `*_ENCODING_MSGPACK_DELTA   = 4`

Backward-compatible: existing producers emitting tags 0-2 see no
change. The vendored proto is the consumer side — DataCollector's
upstream proto can adopt matching enum names in a parallel PR without
breaking either side, since protobuf enums are just integer tags.

## Dispatcher (`drivers/ingest/otel.rs`)

`decode_modified_otlp_sketch_bytes` now matches on the full `encoding`
i32 value rather than only accepting `PROTO = 1`:

  * `PROTO = 1`    → per-sketch `from_sketchlib_proto_bytes`  (existing)
  * `MSGPACK = 3`  → per-sketch `from_msgpack_bytes`          (new)
  * `PROTO_DELTA = 2`, `MSGPACK_DELTA = 4` → deferred, caller falls
                                              through to §5.2 fallback

## Per-sketch decoders

Each of the five accumulators gains a `from_msgpack_bytes(buffer)`
constructor that wraps the sketch-core `deserialize_msgpack` already
present on the underlying type:

  * `CountMinSketchAccumulator::from_msgpack_bytes`
  * `CountSketchAccumulator::from_msgpack_bytes`
  * `DatasketchesKLLAccumulator::from_msgpack_bytes` — and this is
    notable: unlike the `_ENCODING_PROTO` path (which does lossy
    statistical reconstruction via `update()` replay because
    sketchlib's `KllState` proto elides level structure the Rust
    backend types keep private), the msgpack path is a **bit-identical
    round-trip** because sketch-core's `KllSketch` serializes its
    full internal state to msgpack. This gives Go producers a way to
    send KLL sketches losslessly to the Rust backend.
  * `DDSketchAccumulator::from_msgpack_bytes`
  * `HllSketchAccumulator::from_msgpack_bytes`

## Tests

Unit tests — 6 new, 2 per decoder (round-trip + garbage rejection)
on `CountSketchAccumulator`, `DDSketchAccumulator`, and
`HllSketchAccumulator`. CountMin + KLL msgpack paths already had
coverage through `deserialize_from_bytes_arroyo` tests.

E2e test — new `e2e_count_min_sketch_msgpack_modified_otlp_path` in
`tests/e2e_modified_otlp_sketch_path.rs`. Builds a real
`CountMinSketch` in sketch-core, serializes via msgpack, POSTs through
the OTLP HTTP receiver with `encoding = COUNT_MIN_SKETCH_ENCODING_MSGPACK`,
closes a window, and verifies the stored accumulator preserves per-key
counts. Covers the dispatcher path end-to-end for the msgpack branch;
the other four variants share the same branch so per-variant smoke
tests aren't needed for dispatcher coverage (per-variant msgpack
round-trips are covered by the unit tests).

## Drive-by: fix two clippy warnings inherited from main

`cargo clippy --all-targets -- -D warnings` on rust 1.91 flags two
pre-existing issues in the persistence PR #4 code path:
  * `cache.rs` unnecessary `u64 as u64` casts
  * `part.rs` test `&[snap.clone()]` → `std::slice::from_ref(&snap)`

Fixed in this PR since the CI clippy gate would otherwise block PR I.

## Validation

- cargo check --all-targets: clean
- cargo clippy --all-targets -- -D warnings: clean
- cargo fmt --check: clean
- cargo test -p sketch-core --lib: 54 passed
- cargo test -p query_engine_rust --lib: 480 passed (includes 6 new msgpack unit tests)
- cargo test -p query_engine_rust --test e2e_modified_otlp_sketch_path: 6 passed (5 existing PROTO variants + 1 new MSGPACK variant)

+++ asap-query-engine/src/stores/simple_map_store/persistence/cache.rs
@@ -36,7 +36,7 @@ impl PartCache {
                     .weigher(|_k: &PartId, v: &LoadedPart| -> u32 {
                         // Weight = sum of mmap'd bytes. Moka's weigher
                         // returns u32, so clamp large parts.
-                        let len = v.meta.data_len as u64 + v.meta.index_len as u64;
+                        let len = v.meta.data_len + v.meta.index_len;
                         len.min(u32::MAX as u64) as u32
                     })
                     .max_capacity(byte_budget)
